backend/requirements.txt
flask
flask-cors
flask_sqlalchemy
pyodbc
werkzeug
PyJWT